requirements.txt
flask
flask-cors
Pillow
tensorflow
numpy
pymongo
python-dotenv
gunicorn
